### 2018-04-06 #665 @whedon
[PRE REVIEW]: MicroBenthos: a modeling framework for microbial benthic ecology
**Submitting author:** @achennu (<a href="http://orcid.org/0000-0002-0389-5589">Arjun Chennu</a>)
**Repository:** <a href="https://github.com/achennu/microbenthos" target ="_blank">https://github.com/achennu/microbenthos</a>
**Version:** v0.8.2
