### 2018-04-06 #665 @achennu
@arfon Thanks for initiating the review. I just checked the proof. Noticed an error: There is a blank line in the bullet list of features.

@lheagy The software I submitted is useful in biogeochemistry. Based on your topical expertise and that simpeg relates to finite volume estimation, I figured you may be able to easily review my submission.
