### 2018-04-09 #665 @achennu
@lheagy Thanks for taking this on. I don't have other (specific) reviewers to suggest, but I suspect it would be directly within your reach. There is nothing critically specific to chemistry or biology to review. The software essentially is an interface for the user to create partial differential equations. Then these equations are solved, visualized in real time and saved. The benefit for a biogeochem user is that they can work at the abstractions of biomass, microbial group, etc.

There are tutorials in the source tree under `docs/tutorials`, each of which are runnable. While running with `--show-eqns`, the actual constructed PDE is shown. The `--plot` option visualizes the model domain as it happens. From a user perspective, one could try to edit the parameters in the definition file and confirm that these show up in the simulations.

If you could give this a try and let me know if you need more info, then I'd be happy to oblige.
