### 2018-04-09 96a3b53
Renamed Variable to ModelVariable
Moved to core/variable.py
Added DeprecationWarning to old name
Updated tests (passing) in test_model_variable.py
