+++ microbenthos/core/__init__.py
@@ -1,5 +1,6 @@
 from .domain import SedimentDBLDomain
-from .entity import Entity, DomainEntity, Variable
+from .entity import Entity, DomainEntity
+from .variable import ModelVariable, Variable
 from .expression import Expression
 from .irradiance import Irradiance, IrradianceChannel
 from .microbes import MicrobialGroup
