+++ tests/test_model_variable.py
@@ -1,0 +1,304 @@
+import mock
+import pytest
+from fipy import PhysicalField
+from fipy.tools import numerix
+
+from microbenthos import Variable, SedimentDBLDomain
+
+PF = PhysicalField
+
+
+class TestVariable:
+    def test_init_empty(self):
+        with pytest.raises(TypeError):
+            Variable()
+
+    @pytest.mark.parametrize(
+        'unit, err',
+        [('mol', None),
+         ('kg/s', None),
+         ('junk', ValueError),
+         (('kg', 'm'), ValueError),
+         ('m/s', None),
+         ]
+        )
+    def test_create_check_unit(self, unit, err):
+        D = dict(unit=unit)
+
+        if err:
+            with pytest.raises(err):
+                Variable.check_create(**D)
+        else:
+            Variable.check_create(**D)
+
+    def test_create_check_name(self):
+        # supplying name in create params should raise an error
+        with pytest.raises(ValueError):
+            Variable.check_create(name='heh ho')
+
+        Variable.check_create()
+
+    @pytest.mark.parametrize(
+        'constraints, err',
+        [
+            ([('top', 1)], None),
+            ([('bottom', 1)], None),
+            ([('dbl', 1)], None),
+            ([('sediment', 1)], None),
+            ([('top', 0), ('bottom', 1)], None),
+            ([('atop', 0), ('bottom', 1)], ValueError),
+            ([('atop', 0), ('badttom', 1)], ValueError),
+            ([('unknown', 1)], ValueError),
+            (('top', 1), ValueError),
+            (None, ValueError),
+            ('234', ValueError),
+            ([('top', [3, 4])], ValueError),
+            ([('top', 'abc')], ValueError),
+            ([('top', '1.3')], None),
+            ([('top', ['1.3'])], ValueError),
+            ]
+        )
+    def test_constraints_check(self, constraints, err):
+        if err:
+            with pytest.raises(err):
+                Variable.check_constraints(constraints)
+        else:
+            Variable.check_constraints(constraints)
+
+    @pytest.mark.parametrize(
+        'value, unit, hasOld',
+        [
+            (3, 'mol/l', 0),
+            (3.0, 'mol/l', 0),
+            (3.0, 'mol/l', 1),
+            (2.3, None, 0)
+            ]
+        )
+    def test_create_var(self, value, unit, hasOld):
+        # creation casts it into base units
+        create = dict(value=value, unit=unit, hasOld=hasOld)
+        name = 'myVar'
+        v = Variable(name=name, create=create)
+
+        domain = SedimentDBLDomain()
+        v.set_domain(domain)
+        v.setup()
+        assert v.var is domain[name]
+        assert v.var.name == v.name
+        assert v.var.shape == domain.mesh.shape
+        assert (v.var() == PhysicalField(value, unit).inBaseUnits()).all()
+
+    @pytest.mark.parametrize(
+        'constraints',
+        [
+            [('top', PhysicalField('0.2e-3 mol/l'))],
+            [('bottom', PhysicalField('0.2e-3 mol/l'))],
+            [('dbl', PhysicalField('0.2e-3 mol/l'))],
+            [('sediment', PhysicalField('0.4e-3 mol/l'))],
+            [('top', PhysicalField('0.2e-3 mol/l')), ('bottom', PhysicalField(0, 'mol/l'))],
+            [('dbl', PhysicalField('0.4e-3 mol/l')), ('bottom', PhysicalField(0.8e-3, 'mol/l'))],
+            [('sediment', PhysicalField('0.4e-3 mol/l')), ('top', PhysicalField(0.8e-3, 'mol/l'))],
+            # TODO: Figure out a pragmatic policy for these invalid pairs
[…]
