+++ microbenthos/core/variable.py
@@ -1,0 +1,520 @@
+import logging
+
+from fipy import PhysicalField
+from fipy.tools import numerix
+
+from .entity import DomainEntity
+from ..utils.snapshotters import snapshot_var, restore_var
+
+
+class ModelVariable(DomainEntity):
+    """
+    A class to represent a variable on the model domain.
+
+    This class serves as means for defining features such as environmental variables, chemical
+    analytes or microbiological features (biomass). The class allows defining the variable but
+    deferring the initialization of the associated :class:`CellVariable` until the domain is
+    available.
+
+    The interface defined allows to:
+
+        * set boundary condition to :meth:`.constrain` the values
+        * :meth:`.seed` the variable values
+        * create a :meth:`.snapshot` of the state and :meth:`restore_from` it
+        * set :attr:`.clip_min` and :attr:`.clip_max` limits on the values
+
+    """
+
+    def __init__(self, create,
+                 constraints = None,
+                 seed = None,
+                 clip_min = None,
+                 clip_max = None,
+                 **kwargs):
+        """
+        Configure the creation of a variable and its boundary conditions
+
+        Args:
+            name (str): The name of the variable
+
+            create (dict): parameters for variable creation (see :meth:`.create`)
+
+            constraints (dict): Mapping of `location` to value of boundary condition
+                through :meth:`.constrain`
+
+            seed (dict): parameters to seed initial value of variable
+
+        """
+        self.logger = kwargs.get('logger') or logging.getLogger(__name__)
+        kwargs['logger'] = self.logger
+        super(ModelVariable, self).__init__(**kwargs)
+
+        self.logger.debug('Init in {} {!r}'.format(self.__class__.__name__, self.name))
+
+        self.var = None
+        """:type : :class:`fipy.CellVariable` 
+
+        Object created on domain.mesh"""
+
+        self.create_params = self.check_create(**create)
+        """:type : dict
+
+        Validated dict of params for creation of variable"""
+
+        self.logger.debug('{} saving create params {}'.format(self, self.create_params))
+
+        #: mapping of domain location to values for boundary conditions (see :meth:`constrain`)
+        self.constraints = constraints or dict()
+        self.check_constraints(self.constraints)
+
+        self.seed_params = seed or dict()
+
+        #: the min value to which the variable array is clipped
+        self.clip_min = clip_min
+        #: the max value to which the variable array is clipped
+        self.clip_max = clip_max
+
+    def __repr__(self):
+        return 'Var({})'.format(self.name)
+
+    @staticmethod
+    def check_create(**params):
+        """
+        Check that the given `params` are valid for creating the variable once the domain is
+        available
+
+        Args:
+            name (str): a required identifier string
+
+            unit (str): a string like ("kg/m**3") that defines the physical units of the variable
+
+            value (float, :class:`~numpy.ndarray`, :class:`~fipy.PhysicalField`): the value to set
+                for the variable. If a :class:`PhysicalField` is supplied, then its (base) unit is
+                used as `unit` and overrides any supplied `unit`.
+
+        Returns:
+            dict: the params dictionary to be used
+
+        Raises:
+            ValueError: if no `name` is supplied
+
+            ValueError: if `unit` is not a valid input for :class:`PhysicalField`
+
+
+        Note:
+            Due to limitation in :mod:`fipy` (v3.1.3) that meshes do not accept arrays
+            :class:`PhysicalField` as inputs, the variables defined here are cast into base units
+            since the domain mesh is created in meters.
+
+        """
+
+        name = params.get('name')
+        if name:
[…]
