### 2018-04-09 bda04d7
Replaced references to "Variable" by "ModelVariable"
In core/microbes, core/__init__
In model/equation model/model.py
In schema.yml/variables
In test cases (all tests passing)
