### 2018-04-09 5da7468
Updated tutorials
Fixed numerical errors in parameters
Changed code-blocks to literalincludes with highlighting
Fixed mistake in Run it sections
Added a screenshot to concepts.rst
