### 2018-04-12 90d9df1
Bugfix in CLI for overwrite & resume
After last refactor, CLI --resume & --overwrite were not working
Updated logic in Runner now:
--overwrite to delete existing datafile
--resume -1 to resume from existing datafile
Both options respect --confirm
One must be specified if file exists
Resume takes precedence over overwrite
Added CLI option --snapshot-interval
