### 2018-04-12 53de127
Docs fixes: readthedocs github & typos
Added RTD theme options in conf.py (fixes #2)
Clarification + typos in usage.rst
