+++ docs/conf.py
@@ -141,7 +141,13 @@ html_theme = 'sphinx_rtd_theme'
 # Theme options are theme-specific and customize the look and feel of a
 # theme further.  For a list of options available for each theme, see the
 # documentation.
-# html_theme_options = {}
+# FOR READTHEDOCS THEME
+html_theme_options = {
+    'display_version': True,
+    'prev_next_buttons_location': 'both',
+    'sticky_navigation': True,
+    'vcs_pageview_mode': 'blob',
+}
 
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = []
