### 2018-04-12 #674 @achennu
Thanks @leouieda

I have been rerunning the software on some previous model definition files. Found some bugs in the user's CLI interface due to a refactor. I've starting adding commits to the `develop` branch with these minor fixes.

@lheagy Can you clarify about this checkbox?
 Version: Does the release version given match the GitHub release (v0.8.2)?

That version is no longer actual (0.9). I intend that the final version after the review should somehow be referenced?
