### 2018-04-15 achennu/microbenthos#4 @lheagy
Tutorial: suggested improvements
I have worked through the [tutorials in the docs](
https://microbenthos.readthedocs.io/en/latest/tutorials/tut01_define/tutorial.html) as a part of the JOSS Review in openjournals/joss-reviews#674. The text is clear, informative and well-linked to additional information and the API. Congrats!

I just have a couple minor suggestions:
- [x] it would be helpful to have a link to a .yaml file containing the full set of simulation parameters at the top of the file so that users can download scroll through it as they read your tutorial (this also reduces the likely hood of messing up indentations that can easily occur when copy-pasting from a web-page to a local yaml file, particularly if they don't see that a download is available later on in the docs).
- [x] this is minor, but worth addressing: I ran the example on my 13" laptop, and the figure that popped up was larger than the screen (in height), so it might be worth adjusting to also be suitable for smaller screens
