### 2018-04-22 achennu/microbenthos#4 @achennu
@lheagy Thanks for the feedback.

I will move the download links to the definition file to the top of each tutorial. You are right, it would make it more useful for the reader.

As for the size of the graphical exporter, the default is currently set to (1200, 800). I had earlier made it smaller, but I found it hard to read. Instead of trying to guess what other monitors look like, I have exposed the figure size as a user changeable option:

```
microbenthos export video --figsize 8x6 --dpi 100
```
will produce a 800x600 graphical output.

I will dig into whether the ``simulate`` command exposes this option correctly through the ``--exporter`` flag. On second thought, I will look again into making the default smaller.
