### 2018-04-23 b5a117a
Reorganize tutorial + default figure size change
Resolves issue #4
