### 2018-04-23 achennu/microbenthos#4 @achennu
@lheagy I've now updated the tutorials to contain a download link to the definition at the top. Good idea.

I have also changed the default size of the plotter to (800, 600). This should be small enough for most screens, and anyhow the window can be maximized without trouble.
