### 2018-04-23 #674 @achennu
@lheagy @leouieda
Thanks for your inputs on the repo. I've made the changes you've indicated thus far. Travis builds are passing. Once you indicate that the JOSS review is through, I'll bump version, release and create an archive on zenodo.
