### 2018-04-24 achennu/microbenthos#4 @achennu
The docs haven't updated yet. It is an upstream issue with readthedocs https://github.com/rtfd/readthedocs.org/issues/3998

But you can can see the last commit on develop to see the change.
