### 2018-04-25 achennu/microbenthos#4 @achennu
@lheagy The problem with readthedocs seems to be fixed.
The updated version can be seen at: http://microbenthos.readthedocs.io/en/develop/tutorials/tut01_define/tutorial.html
