### 2018-05-01 achennu/microbenthos#5 @leouieda
Ouput definition file overwriting the input file
@achennu in the [Usage docs](https://microbenthos.readthedocs.io/en/develop/usage.html#running-simulations) there is a note saying:

> The default name for the output definition file is definition.yml. So if your input file is called definition.yml, then it will get overwritten on running the command.

Wouldn't it be better to check if the input and output names are the same and raise an exception instead? Or ask the user if that's OK? I see that there are confirmation messages for overwriting the simulation HDF5 output.

Related to openjournals/joss-reviews#674
