### 2018-05-01 achennu/microbenthos#6 @leouieda
Minor correction to the paper
Missing verb in "MicroBenthos is a modeling framework to study in silico microbenthic habitats." Something along the lines of "framework *used* to study" or "framework *created* to"?

openjournals/joss-reviews#674
