### 2018-05-01 #674 @lheagy
Thanks @leouieda! In your opinion, should these issues be addressed before the paper is published?
