### 2018-05-02 achennu/microbenthos#5 @achennu
@leouieda that's well spotted. However, it is deliberately so. The idea is that the definition file supplied by the user goes through a validation. At this stage, the definition dictionary becomes fully explicit, i.e. the implicit and default parameters get filled out, and so the written ``definition.yml`` file is complete.
So what could happen is:
* The user's input file does not get validated, in which case nothing is written to disk. Problems in the definition file are listed on the console, so the user can correct them.
* The user can check if the input they provide can be validated by using the ``microbenthos export model`` command.
* If the definition file gets validated, then the validated dictionary is written to ``definition.yml``. In this step, no information is lost (only enhanced).

The ``definition.yml`` file is the starting point for the simulations. On running one, a ``runner.yml`` file captures the library versions used in the runtime. Then when data is produced, it gets stored into the ``simulation_data.h5`` store. With this setup, having the definition file and the data file is enough to resume a simulation. Technically, all the information for the simulation is already stored in the metadata of the hdf file, but there are additional steps I have implemented to ensure that a model cannot be 'wrongly' resumed leading to spurious data. I had considered to include docs about this, but I'd need some more testing to ensure it as a guarantee.

As such, the checks for overwriting are not performed for the defnition or runner files, but mainly for the data file only -- which can either be resume, or overwrite.
