### 2018-05-02 #674 @achennu
Thanks @leouieda ! I'm glad you enjoyed it.

The ``export model`` command is used to validate an input file by the user. That command writes out the validated YAML string to stdout, which can of course be piped to a file for further use.

HDF5 is a widely used standard format for storing scientific data. It contains many useful features: it is self-documenting, supports structured metadata, can be resized, etc. The specifications for it are codified by the HDF group. I use the vanilla h5py library to read and write to this format, but there are other libraries that build on it like pytables. There are of course readers in matlab, R, C++, etc. In fact, there is the java based GUI called ``hdfview`` to inspect it, though I'm partial to `vitables` written in python. With these, one can double click the file, inspects the hierarchy, the nodes, the metadata attached, etc. For microbenthos, the HDF5 format serves as a good archival output target. It is not tied down to the microbenthos software, or even python for that matter. The produced data can be consumed in whatever way serves the user's needs best: export to video, export to frames, open as interactive visualization, or read and analyze in their tool of choice.

I hope this satisfies all your concerns. Thanks for taking the time and giving useful feedback.
