### 2018-05-02 7968c5d
Added JOSS banner and small fix in paper text
