### 2018-05-02 achennu/microbenthos#5 @leouieda
OK, so `definition.yml` is a completed version of the input file. That makes sense. It's a smart way to preserve saved simulations from changes in default parameters in future releases.

Have you considered using `tempfile.NamedTemporaryFile` to generate a unique ID for each simulation? You can append this ID to the end of every file name so you'd have `simulation_data_ID.h5` and `definition_ID.yml`. They can still know of each other through the ID and  repeated runs won't cause files to be overwritten. I use this in [GMT/Python to generate unique figure names](https://github.com/GenericMappingTools/gmt-python/blob/26c4849706f233a01854e727b030dfedb6c25efd/gmt/helpers/tempfile.py#L10).
