+++ docs/conf.py
@@ -87,9 +87,9 @@ copyright = u"2018, Arjun Chennu"
 # the built documents.
 #
 # The short X.Y version.
-version = "0.9"
+version = "0.9.1"
 # The full version, including alpha/beta/rc tags.
-release = "0.9"
+release = "0.9.1"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
