### 2018-05-04 achennu/microbenthos#5 @achennu
I did start with a unique id based tagging of the simulation definition and data files. However, the point of modelling is to compare the changes in the parameters on some metric. Running a family of simulations, I soon realized that it becomes difficult to keep track of which run has which parameters based on just a random unique id. I instead developed a rig to run each simulation in parallel and the natural output of a run is a separate directory. Easier to parse, navigate and prune. In a future version, I may include this parallel running setup into the repo.

Will close this issue for now. Thanks for the input!
